pyproject.toml
[project]
name = "policy-governed-agent-runtime"
version = "0.1.0"
description = "Policy-governed AI agent runtime with economics and audit controls."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "pydantic>=2.8.0",
  "PyYAML>=6.0.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
